% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_exon_loci}
\alias{add_exon_loci}
\title{Add exon loci}
\usage{
add_exon_loci(splice_data)
}
\arguments{
\item{splice_data}{dataframe that contains a column named splice_pattern}
}
\value{
original dataframe with additional columns (and therefore rows) for
start and end of each exon in the splice pattern
}
\description{
parses splice patterns, adding 2 columns (start and end of exon) to a dataframe
containin the splice patterns
}
\examples{
none
}
