% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{add_unknown_ids}
\alias{add_unknown_ids}
\title{Add more useful ids to unknown transcripts.
Append a number to any Transcript_ids labelled with the id "unknown".}
\usage{
add_unknown_ids(
  parsed_splices,
  Transcript_id_col = "Transcript_id",
  variant_col = "variant"
)
}
\arguments{
\item{parsed_splices}{dataframe containing columns named variant, strand, score,
Transcript_id and splice_pattern. Intended to be the output of parse_nexons_gtf
function.}

\item{Transcript_id_col}{name of the Transcript id column}

\item{variant_col}{name of column containing the variant id}
}
\value{
tibble
}
\description{
Add more useful ids to unknown transcripts.
Append a number to any Transcript_ids labelled with the id "unknown".
}
\examples{
NA
}
