% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkJunctionMatches}
\alias{checkJunctionMatches}
\title{Check junction matches
Check if 2 sets of splice junctions match.}
\usage{
checkJunctionMatches(set1, set2, flexibility = 0)
}
\arguments{
\item{set1}{tibble with column names junction_no, start, end}

\item{set2}{tibble with column names junction_no, start, end}

\item{flexibility}{integer value - number of bases away to still be classified as a match. For exact matches set flexibility to 0. Default: 0}
}
\value{
vector of TRUE and FALSE values
}
\description{
Check junction matches
Check if 2 sets of splice junctions match.
}
\examples{
NA
}
