% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkTruncation}
\alias{checkTruncation}
\title{checkTruncation}
\usage{
checkTruncation(other_set, potential_truncs, flexibility = 0)
}
\arguments{
\item{other_set}{tibble - one set of splice junctions}

\item{potential_truncs}{tibble - one set of splice junctions to test to see if they might be truncations}

\item{flexibility}{integer value - number of bases away to still be classified as a match. For exact matches set flexibility to 0. Default: 0}
}
\value{
TRUE or FALSE value
}
\description{
Checks one set of splice junctions against another set, to see if one is a potential truncation of the other.
Does an initial check for the first splice junction location, then do an exact test (within flexibility parameter) for vectors of the same length.
}
\examples{
NA
}
