% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identifyPotentialTruncations}
\alias{identifyPotentialTruncations}
\title{identifyPotentialTruncations}
\usage{
identifyPotentialTruncations(parsed_splices, flexibility = 10)
}
\arguments{
\item{parsed_splices}{tibble containing set of splices - output from \code{\link{parse_nexons_gtf}}}

\item{flexibility}{integer value - number of bases away to still be classified as a match. For exact matches set flexibility to 0. Default: 10}
}
\value{
modified tibble with additional column named truncation_origin
}
\description{
Identifies whether transcripts might be truncations of other transcripts.
}
\examples{
file <- system.file("extdata", "inst/extdata/nexons_sirv5_f15_trunc.txt", package = "nexonsAnalysis")
nexons_output <- readr::read_delim(file)
parsed_splices <- parse_nexons_gtf(nexons_output, min_count = 3)
parsed_with_trunc <- identifyPotentialTruncations(parsed_splices)

}
