% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_default_nexons}
\alias{parse_default_nexons}
\title{Parse default nexons output}
\usage{
parse_default_nexons(nexons_output, score_column, min_count = 1)
}
\arguments{
\item{nexons_output}{dataframe containing columns named "Gene ID", "Strand", "SplicePattern", "Transcript id" and one containing the scores, the name of which is specified in the score_column argument}

\item{score_column}{name of the column containing the scores}

\item{min_count}{minimum count for the splice pattern to be included (default: 1)}
}
\value{
tibble containing the columns "variant", "strand", "score", "Transcript_id", "Gene_id", "splice_pattern"
}
\description{
Takes a default nexons output file (not a gtf - see parse_nexons_gtf) and makes it
compatible, so that it can be used as input to draw_splice_picture
}
\examples{
file <- system.file("extdata", "sirv5.txt", package = "nexonsAnalysis")
nexons_output <- readr::read_delim(file)
parsed_splices <- parse_default_nexons(nexons_output, score_column = "seqs_sirv5_minimap.sam")
}
