% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_nexons_gtf}
\alias{parse_nexons_gtf}
\title{Parse gtf output from nexons}
\usage{
parse_nexons_gtf(nexon_gtf, min_count = 1)
}
\arguments{
\item{nexon_gtf}{gtf file generated by nexons}

\item{min_count}{minimum count for transcript to be displayed (default: 1)}
}
\value{
tibble containing the columns "variant", "strand", "score", "Transcript_id", "Gene_id", "splice_pattern"
}
\description{
Takes a gtf file generated by nexons and separates the attribute column
into Transcript_id, Gene_id and splice_pattern
}
\examples{
file <- system.file("extdata", "inst/extdata/nexons_sirv5_f15.gtf", package = "nexonsAnalysis")
nexons_output <- readr::read_tsv(file)
parsed_splices <- parse_nexons_gtf(nexons_output)

}
