% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_quant}
\alias{plot_quant}
\title{Quantitative plot of transcripts}
\usage{
plot_quant(splice_df)
}
\arguments{
\item{splice_df}{same as the input to draw_splice_picture, a data frame of
transcripts. Must contain columns named variant, score and Transcript_id.}
}
\value{
ggplot object
}
\description{
Quantitative plot of transcripts
}
\examples{
NA
}
