% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_wrapper}
\alias{plot_wrapper}
\title{Plot wrapper}
\usage{
plot_wrapper(nexons_output_file, quant_plot = FALSE, min_count = 1)
}
\arguments{
\item{nexons_output_file}{}
}
\value{

}
\description{
Wrapper around splice plotting function for a nexons output file.
}
\examples{
file <- system.file("extdata", "nexons_sirv5_f15.gtf", package = "nexonsAnalysis")
# just the splice picture
plot_wrapper(file, min_count = 5)
# including a quantitative plot
plot_wrapper(file, quant_plot = TRUE, min_count = 5)
}
