% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spliceStartsEnds}
\alias{spliceStartsEnds}
\title{spliceStartsEnds
Takes a splice string in the format "8381:8455-8585:10859-10991:11312" and converts it to a tibble of start and end points of splice junctions.
Used internally by \code{\link{identifyPotentialTruncations}} function.}
\usage{
spliceStartsEnds(splice_string)
}
\arguments{
\item{splice_string}{set of splices in the format 8381:8455-8585:10859-10991:11312"}
}
\value{
tibble of start and end points of splice junctions
}
\description{
spliceStartsEnds
Takes a splice string in the format "8381:8455-8585:10859-10991:11312" and converts it to a tibble of start and end points of splice junctions.
Used internally by \code{\link{identifyPotentialTruncations}} function.
}
\examples{
splice <- "8381:8455-8585:10859-10991:11312"
spliceStartsEnds(splice)
}
