% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{draw_splice_picture}
\alias{draw_splice_picture}
\title{Draw a picture of splice patterns}
\usage{
draw_splice_picture(splice_data, title_text = "", quant = FALSE)
}
\arguments{
\item{splice_data}{dataframe containing columns named variant, strand, score,
Transcript_id and splice_pattern. Intended to be the output of parse_nexons_gtf
function.}

\item{title_text}{optional title for plot (probably gene name)}

\item{quant}{TRUE or FALSE whether to show a quantitative plot of transcipt
counts alongside the splice variants}
}
\value{
ggplot object
}
\description{
Creates plot showing splice variants
}
\examples{
NA
}
