% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_wrapper}
\alias{plot_wrapper}
\title{Plot wrapper}
\usage{
plot_wrapper(
  nexons_output_file,
  order_splices = NULL,
  quant_plot = FALSE,
  min_count = 1
)
}
\arguments{
\item{nexons_output_file}{gtf file output from nexons}

\item{order_splices}{(one of 'score', 'name', NULL) default NULL. How to order the splices on the y axis.}

\item{quant_plot}{TRUE or FALSE, whether to display a quantitative plot
alongside the splice picture}

\item{min_count}{minimum number of times the splice variant must be present to
be display. Any splice variants with counts < min_count will be filtered out.}
}
\value{
splice plot
}
\description{
Wrapper around splice plotting function for a nexons output file.
}
\examples{
file <- system.file("extdata", "nexons_sirv5_f15.gtf", package = "nexonsAnalysis")
# just the splice picture
plot_wrapper(file, min_count = 5)
plot_wrapper(file, min_count = 5, order_splices = "name")
# including a quantitative plot
plot_wrapper(file, quant_plot = TRUE, min_count = 5, order_splices = "score")
}
